fix: cast bigint to integer in V17 date arithmetic
V17 used CURRENT_DATE + (id % 14 + 1) where id is bigint;
PostgreSQL has no date + bigint operator, only date + integer.

+++ backend/src/main/resources/db/migration/V17__update_test_dates.sql
@@ -1,5 +1,5 @@
 -- Update test dates so upcoming tests appear in student dashboard
 -- Spread ~2/3 of past tests across the next 14 days so the "Upcoming Tests" section shows data
 UPDATE tests
-SET test_date = CURRENT_DATE + (id % 14 + 1)
+SET test_date = CURRENT_DATE + ((id % 14 + 1)::integer)
 WHERE test_date < CURRENT_DATE AND id % 3 != 0;
